refactor: consts/rank.tsを作成した #256
